requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
matplotlib>=3.5.0
streamlit>=1.28.0
Pillow>=10.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.3
